Rename the input arguments of hutchinson.hutchinson
